+++ elfo-logger/src/actor.rs
@@ -1,4 +1,4 @@
-use std::sync::Arc;
+use std::{sync::Arc, time::Duration};
 
 use metrics::increment_counter;
 use tokio::{
@@ -12,7 +12,7 @@ use elfo_core::{
     messages::{ConfigUpdated, Terminate},
     msg,
     signal::{Signal, SignalKind},
-    ActorGroup, Blueprint, Context, TerminationPolicy,
+    ActorGroup, Blueprint, Context, RestartParams, RestartPolicy, TerminationPolicy,
 };
 
 use crate::{
@@ -42,6 +42,10 @@ impl Logger {
         ActorGroup::new()
             .config::<Config>()
             .termination_policy(TerminationPolicy::manually())
+            .restart_policy(RestartPolicy::on_failure(RestartParams::new(
+                Duration::from_secs(5),
+                Duration::from_secs(30),
+            )))
             .stop_order(105)
             .exec(move |ctx| Logger::new(ctx, shared.clone(), filtering_layer.clone()).main())
     }