+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - utils: slightly improve performance of `RateLimiter`.
 - telemeter: handle only `GET /metrics` requests.
 - telemeter: rename the `address` config parameter to `listen` with alias to the old name.
+- **BREAKING** restarting: The default restart policy has been updated. `RestartPolicy::never()` is now set by default ([#118]).
+- restarting: linear backoff is replaced with exponential backoff ([#118]).
+- restarting: the restart policy can be overridden through configuration ([#118]).
+- context: added `Context::start_info` to retrieves information related to the start of the actor ([#118]).
+
+[#118]: https://github.com/elfo-rs/elfo/pull/118
 
 ### Fixed
 - telemeter: avoid spawning tasks for a HTTP server, it improves common metrics.